Move add command to AddCommand
