+++ tests/conftest.py
@@ -1,11 +1,14 @@
 # -*- coding: utf-8 -*-
 
+import os
 import shutil
 import tempfile
 
 import boto3
 
 from faker import Faker
+
+import mock
 
 import moto
 
@@ -15,6 +18,17 @@ from s4.clients.local import LocalSyncClient
 from s4.clients.s3 import S3SyncClient
 
 fake = Faker()
+
+
+@pytest.yield_fixture
+def config_file():
+    fd, temp_path = tempfile.mkstemp()
+    mocker = mock.patch('s4.utils.CONFIG_FILE_PATH', temp_path)
+    mocker.start()
+    yield temp_path
+    mocker.stop()
+    os.close(fd)
+    os.remove(temp_path)
 
 
 @pytest.yield_fixture
