+++ tests/test_utils.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 
+import json
+
 import mock
 
 from s4 import utils
@@ -21,3 +23,15 @@ def test_get_input_secret(getpass, input_fn):
 
     assert getpass.call_count == 0
     assert input_fn.call_count == 1
+
+
+class TestGetConfigFile(object):
+    @mock.patch('s4.utils.CONFIG_FILE_PATH', '/i/dont/exist')
+    def test_no_file(self):
+        assert utils.get_config() == {'targets': {}}
+
+    def test_correct_output(self, config_file):
+        with open(config_file, 'w') as fp:
+            json.dump({'local_folder': '/home/someone/something'}, fp)
+
+        assert utils.get_config() == {'local_folder': '/home/someone/something'}
